Separate code and settings
