+++ settings/questionsSelection.ts
@@ -1,0 +1,55 @@
+/**
+ * This file contains a list of all questions available on the give instance of SEAL. 
+ * TODO: In future this file should be generated by a script.
+ */
+import { QuestionRoutes } from "../shared/src/api/QuestionRouter"
+
+import { TestQuestion } from "../shared/src/question-generators/test/test"
+import { Between } from "../shared/src/question-generators/asymptotics/between"
+import { LandauNotation } from "../shared/src/question-generators/asymptotics/landau"
+import { AsymptoticsPreciseLanguage } from "../shared/src/question-generators/asymptotics/precise-language"
+import { SortTerms } from "../shared/src/question-generators/asymptotics/sort"
+import { SimplifySum } from "../shared/src/question-generators/asymptotics/sum"
+import { RecursionFormula } from "../shared/src/question-generators/recursion/formula"
+import { RecurrenceMaster } from "../shared/src/question-generators/recursion/recurrenceMaster"
+import { Loops } from "../shared/src/question-generators/time/loops"
+
+/** List of all question routes */
+export const allQuestionGeneratorRoutes: QuestionRoutes = [
+  {
+    path: "test/test",
+    generator: TestQuestion,
+  },
+  {
+    path: "asymptotics/precise-language",
+    generator: AsymptoticsPreciseLanguage,
+  },
+  {
+    path: "asymptotics/sort",
+    generator: SortTerms,
+  },
+  {
+    path: "asymptotics/landau",
+    generator: LandauNotation,
+  },
+  {
+    path: "asymptotics/sum",
+    generator: SimplifySum,
+  },
+  {
+    path: "asymptotics/between",
+    generator: Between,
+  },
+  {
+    path: "recursion/formula",
+    generator: RecursionFormula,
+  },
+  {
+    path: "recursion/master",
+    generator: RecurrenceMaster,
+  },
+  {
+    path: "time/loops",
+    generator: Loops,
+  },
+]
